Refactoring

+++ .zsh/shared.aliases.zsh
@@ -1,6 +1,8 @@
 # This file is for shared aliases, ones that are useful
 # in a no-project session as well as in a project one.
 
+save-function-list
+
 # Core.
 alias df='df -h'
 alias ls='ls -F'
@@ -28,37 +30,39 @@ du() {
   done
 }
 
+report-custom-functions "Aliases" && save-function-list
+
 # Emacs.
 verify-emacs-session() {
-  if test -S /tmp/emacs$(id -u)/$1; then
-    return 0
-  else
-    echo "No such session: $1"
-    return 1
-  fi
+  [[ -v 1 ]] || error "Usage: verify-emacs-session session-name" && return 1
+  ls-emacs-sessions | grep $1 && return 0
+  error "No such session: $1" && return 1
 }
 
 verify-absence-of-emacs-session() {
-  if ! test -S /tmp/emacs$(id -u)/$1; then
-    echo "$(tput setaf 2)~$(tput sgr0) Starting Emacs session $(tput setaf 7)$1$(tput sgr0)"
-  fi
+  [[ -v 1 ]] || error "Usage: verify-absence-of-emacs-session session-name" && return 1
+  ls-emacs-sessions | grep $1 && return 1
 }
 
-running-emacs-sessions() {
+ls-emacs-sessions() {
   ps aux | egrep "emacs --daemon=\w+" | grep -v grep | awk '{ print $12 }' | sed -E 's/.*--daemon=(\w+)/\1/'
 }
 
 stop-all-emacs-sessions() {
-  for session in $(running-emacs-sessions); do
+  for session in $(ls-emacs-sessions); do
     stop-emacs-session $session
   done
 }
 
 start-emacs-session() {
-  verify-absence-of-emacs-session && emacs --daemon=$1
+  [[ -v 1 ]] || error "Usage: start-emacs-session session-name" && return 1
+  verify-absence-of-emacs-session && rm /tmp/emacs$(id -u)/$1 && emacs --daemon=$1
 }
 
 stop-emacs-session() {
+  [[ -v 1 ]] || error "Usage: stop-emacs-session session-name" && return 1
   verify-emacs-session $1 || return 1
   kill $(ps aux | egrep "emacs --daemon=$1" | awk '{ print $2  }')
 }
+
+report-custom-functions "Emacs" && echo
